Remove width specific settings for table view
We are using the `terminal-table` gem to display the table view,
but we were explicitly setting up the width for the table which
was causing some issue when our content exceed that limit.

This commit fixes that issue and left the width settings to the
gem default which should be the full width of the screen.

Fixes #10