Add interface to list files for a Space
In Riobse, the files are organized based on a user space, and this
commit adds the interface to retrieve the list of files. By default
this will return the basic details in tabular format.

But we also have an one additional option `--format` that allows us
to specify the output format, current accepted option is `json`.

```sh
ribose file list --space-id 123456
```