Add interface to retrieve space members
The ribose client offers `Ribose::Member` interface that allows us
to retrieve the list of members for any specific space. This commit
usage that interface and provides a command line interface for that
by default it shows some basic data in tabular format, but we can
customize that by using the `format` option.

```sh
ribose member list --space-id space_uuid
```