Add interface to add new space member
Ribose API offers `Ribose:SpaceInvitation` interface that allow us
to invite a user to a space, and this commit usages that interface
and provides a CLI interface for that. The normal use case is bit
different than the usual command, so please feel free to run the
`ribose member add --help` if you need any guideline.

```sh
ribose member add \
  --space-id space_uuid \
  --user-id=user-one-uuid:role_one_id user-two-uuid:role_two_id \
  --email=email-one@example.com:role_one_id email@example.com:role_id \
  --message="Your invitation messages to the invitees"
```