Add interface to list space invitations
Ribose API offers `Ribose::SpaceInvitation`, which allows us to
retrieve the list of space invitation for the current user, and
this commit adds the command line interface for that. By default
it will list the details in tabular format, but we can customize
it using the `format` option.

This interface also supports `query` option, which we can use to
pass customer query parameters as `key:value` pair.

```sh
ribose invitation list --space-id 1234 [--query=key:value]
```