Add interface to list join space requests
This commit adds a CLI interface to list the join space requests, by
default it will list some basic attributes in the tabular format. But
if we need all the attributes then we pass `json` to  `--format` and
it will list all attributes as JSON.

This interface also support one additional hash arguments `query`,
which allows us to pass any attributes that will be passed along as
a query parameters.

```sh
ribose join-space list [--query=space-id:2468]
```