Add interface to add join space request
This commit adds an interface to create a new join space request.
The `space_id` is the required parameter for this interface, but
this also support additional option for `type`, `state` and the
`message`. If none of the optional parameters are provided then
it also sets some standard default for this interface.

```sh
ribose join-space add --space-id 1234 [--message "My request message"]
```