Add interface to fetch file details
This commit adds the `show` interface for `file` command. It expect
us to provide the `--file-id` and `--space-id` option then it will
print out the details in  tabular format. If we need a different
output format then we can use the `--format` option for that.

```sh
ribose file show --file-id 5678 --space-id 1234 [--format json]
```